# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SimEngBay API",
    "version": "baseline"
  },
  "paths": {
...
    "/api/simulations": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Create",
        "description": "형상(STEP)과, 구속을 걸 거라면 CAD 가 보낸 `topology.json` 을 함께 받는다.",
        "operationId": "create_api_simulations_post",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_api_simulations_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "List Simulations",
        "operationId": "list_simulations_api_simulations_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "workspace_slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Slug"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SimulationSummaryOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/doe/preview": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Preview Doe",
        "description": "DOE 폴더를 훑어 본다 — **걸기 전에** 점 몇 개, 변수 무엇, 건너뛸 것 몇 개인지.\n\n폴더는 **서버가 보는 경로**다(공유 스토리지). 브라우저가 파일을 올리는 것이 아니다 —\n설계점 200개면 STEP 만 수십 MB 다.",
        "operationId": "preview_doe_api_simulations_doe_preview_get",
        "parameters": [
          {
            "name": "path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Path"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoePreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/doe/import": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Import Doe",
        "description": "폴더 한 벌을 해석 작업 N 개로. 걸 수 없는 점은 **이유와 함께** 돌려준다.",
        "operationId": "import_doe_api_simulations_doe_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DoeImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoeImportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Get One",
        "operationId": "get_one_api_simulations__simulation_id__get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/status": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Poll Status",
        "description": "상세와 같다. **폴링이 쓰는 경로라 접근 로그에 안 남는다.**",
        "operationId": "poll_status_api_simulations__simulation_id__status_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/result": {
      "get": {
        "tags": [
          "simulations"
        ],
        "summary": "Get Result",
        "description": "`result.json` 그대로 — 모드 목록 · 단위계 · 참여계수 · 모드 형상 파일 이름.\n\n**모양을 스키마로 굳히지 않는다.** 레시피마다 다르고(모달 · 정적), 해석이 낸 파일이\n정본이다. 화면은 필요한 칸만 읽는다.",
        "operationId": "get_result_api_simulations__simulation_id__result_get",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Result Api Simulations  Simulation Id  Result Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/simulations/{simulation_id}/retry": {
      "post": {
        "tags": [
          "simulations"
        ],
        "summary": "Retry",
        "operationId": "retry_api_simulations__simulation_id__retry_post",
        "parameters": [
          {
            "name": "simulation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Simulation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ArtifactOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage",
          "kind",
          "filename",
          "content_type",
          "size_bytes",
          "created_at"
        ],
        "title": "ArtifactOut"
      },
...
      "Body_create_api_simulations_post": {
        "properties": {
          "spec": {
            "type": "string",
            "title": "Spec",
            "description": "JobSpec JSON (app/core/spec.py)"
          },
          "workspace_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Slug"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "topology": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Topology"
          }
        },
        "type": "object",
        "required": [
          "spec",
          "file"
        ],
        "title": "Body_create_api_simulations_post"
      },
...
      "DoeImportOut": {
        "properties": {
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Created"
          },
          "skipped": {
            "items": {
              "$ref": "#/components/schemas/DoePointPreview"
            },
            "type": "array",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "study_id",
          "name",
          "created",
          "skipped"
        ],
        "title": "DoeImportOut"
      },
      "DoeImportRequest": {
        "properties": {
          "path": {
            "type": "string",
            "minLength": 1,
            "title": "Path"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "workspace_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Slug"
          },
          "numbers": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numbers"
          }
        },
        "type": "object",
        "required": [
          "path",
          "spec"
        ],
        "title": "DoeImportRequest"
      },
      "DoePointPreview": {
        "properties": {
          "number": {
            "type": "integer",
            "title": "Number"
          },
          "params": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Params"
          },
          "usable": {
            "type": "boolean",
            "title": "Usable"
          },
          "skip_reason": {
            "type": "string",
            "title": "Skip Reason",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "number",
          "params",
          "usable"
        ],
        "title": "DoePointPreview",
        "description": "가져오기 전에 보여 줄 한 줄 — **걸 수 있나, 아니면 왜 못 거나.**"
      },
      "DoePreviewOut": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "factors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Factors"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "default": ""
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/DoePointPreview"
            },
            "type": "array",
            "title": "Points"
          },
          "usable": {
            "type": "integer",
            "title": "Usable"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "path",
          "study_id",
          "name",
          "factors",
          "points",
          "usable",
          "skipped"
        ],
        "title": "DoePreviewOut",
        "description": "폴더를 훑어 본 결과.\n\n**먼저 보여 주고 나서 건다** — 200개를 잘못 걸면 되돌리기 어렵다."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_SimulationSummaryOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SimulationSummaryOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[SimulationSummaryOut]"
      },
...
      "SimulationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "recipe": {
            "type": "string",
            "title": "Recipe"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "source_ref": {
            "type": "string",
            "title": "Source Ref"
          },
          "source_meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Meta"
          },
          "owner_workspace_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Workspace Id"
          },
          "owner_workspace_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Workspace Name"
          },
          "requested_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Id"
          },
          "requested_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Name"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "spec": {
            "additionalProperties": true,
            "type": "object",
            "title": "Spec"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageOut"
            },
            "type": "array",
            "title": "Stages"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactOut"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts"
          },
          "worker_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Worker Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "recipe",
          "status",
          "source_kind",
          "source_ref",
          "source_meta",
          "owner_workspace_id",
          "owner_workspace_name",
          "requested_by_id",
          "requested_by_name",
          "error_code",
          "summary",
          "created_at",
          "started_at",
          "finished_at",
          "spec",
          "stages",
          "error_message",
          "artifacts",
          "attempts",
          "worker_id"
        ],
        "title": "SimulationOut"
      },
      "SimulationSummaryOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "recipe": {
            "type": "string",
            "title": "Recipe"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "source_kind": {
            "type": "string",
            "title": "Source Kind"
          },
          "source_ref": {
            "type": "string",
            "title": "Source Ref"
          },
          "source_meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Source Meta"
          },
          "owner_workspace_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Workspace Id"
          },
          "owner_workspace_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Workspace Name"
          },
          "requested_by_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Id"
          },
          "requested_by_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By Name"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "recipe",
          "status",
          "source_kind",
          "source_ref",
          "source_meta",
          "owner_workspace_id",
          "owner_workspace_name",
          "requested_by_id",
          "requested_by_name",
          "error_code",
          "summary",
          "created_at",
          "started_at",
          "finished_at"
        ],
        "title": "SimulationSummaryOut",
        "description": "목록 한 줄. 스펙 · 단계 · 산출물은 싣지 않는다 — 50줄에 그것까지 실으면 목록이\n무겁다."
      },
      "StageOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "type": "object",
        "required": [
          "name",
          "status"
        ],
        "title": "StageOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}